+++ familyhelper-clannad-sdk/src/main/java/com/dwarfeng/familyhelper/clannad/sdk/util/Constraints.java
@@ -44,9 +44,9 @@ public final class Constraints {
     public static final int LENGTH_USER = 50;
 
     /**
-     * 消息的长度约束。
+     * 主题的长度约束。
      */
-    public static final int LENGTH_MESSAGE = 255;
+    public static final int LENGTH_SUBJECT = 100;
 
     private Constraints() {
         throw new IllegalStateException("禁止实例化");